tui/requirements.txt
textual-dev
requests